requirements.txt
discord.py==2.3.2
psycopg2==2.9.9
requests==2.31.0
aiohttp==3.9.1